fix(test): cast getTotalLoadedClassCount() result to Long
ClassLoadingMXBean.getTotalLoadedClassCount() returns a long, so the reflective
invoke returns a boxed Long. The previous `as Int` cast threw ClassCastException
in the forked JVM during warmup, so the benchmark printed no CLASSES_LOADED and
the test failed identically on both branches at the "Subprocess emitted no
CLASSES_LOADED" guard. Cast to Long.

+++ sdk/core/src/test/java/com/klaviyo/core/benchmark/RegistryStartupBenchmark.kt
@@ -69,5 +69,5 @@ private fun totalLoadedClasses(): Long {
     val managementFactory = Class.forName("java.lang.management.ManagementFactory")
     val bean = managementFactory.getMethod("getClassLoadingMXBean").invoke(null)
     val beanInterface = Class.forName("java.lang.management.ClassLoadingMXBean")
-    return (beanInterface.getMethod("getTotalLoadedClassCount").invoke(bean) as Int).toLong()
+    return beanInterface.getMethod("getTotalLoadedClassCount").invoke(bean) as Long
 }
